+++ changelog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 <!-- Add all new changes here. They will be moved under a version at release -->
 * `FIX` Fixed an issue preventing to set the locale to Japanese
+* `FIX` Handle alternate linux config directory paths for addons
 
 ## 3.11.0
 * `NEW` Added support for Japanese locale